Refactoring to separate responsibilities more
